// src/components/types/urlParam.ts
export enum UrlParam {
    PAGE = 'page',
    PER_PAGE = 'perPage',
    SORT_BY = 'sortBy',
    SWITCH_VIEW = 'switchView',
    SEARCH_QUERY = 'q',
    MIN_PRICE = 'minPrice',
    MAX_PRICE = 'maxPrice',
    MIN_SIZE = 'minSize',
    MAX_SIZE = 'maxSize',
    MIN_STOCK = 'minStock',
    MAX_STOCK = 'maxStock',
    COLORS = 'colors',
    COLLECTIONS = 'collections',
    CATEGORIES = 'categories',
    SHOW_FILTERS = 'showFilters',
    BUY_NOW = 'buyNow',
}